Add a zero argument break and return and make the expression type checking consistent with the fall-through expression.
The prior code make the number of arguments to break and return operators dependent on the number of expected values, however this property does not hold in general. For example consider a single break with a single expression that returns the type (), the same type as `nop`, here the number of expected values could be zero yet there is one argument.

Further the prior code appeared to make an interpretation of the arguments as a tuple constructor that would increase with multiple-value support, yet this is not consistent with the fall-through which is a single expression.

The function return operator has the same issue. Even though the function expected results might be considered side-information, the number of arguments to return does not in general match the number of expected function results.

Wasm currently has the start of multiple value expressions, there are single value results and the empty-values result from nop, so even now this needs some consideration to be consistent.

There is some interest in exploring serializations of the AST that are not dependent on the context, and thus a need to have each operator have a defined number of arguments, not necessarily fixed but if not fixed then encoding in the operator.

There is also a strong interest in having an efficient encoding for the zero argument break and return operators, even if this means exposing this at the AST level.

The solution here adds separate zero expression break and return operators, which have a fixed number of arguments. These are only valid when the expected number of values is zero, but this does not add a new constraint and is still consistent with the fall-through expressions. The current break and return operators become fixed argument single expression operators for which their expression has the same semantics as the fall-through.

Future multi-value expression support would be expected to add an explicit tuple or values constructor, matters yet to be decided, and these would likely be usable for the fall-through and the break or return expression so no new break or return expressions would be needed.