Put br_if conditions last.
This implements the second half of https://github.com/WebAssembly/design/pull/489.

Desugaring br_if to if+break was fairly awkward; adding br_if to the kernel
language was straightforward and avoided the awkwardness.

This change happened to expose an aspect of `br_if`'s type checking which
differed from my expectations, so I've taken the liberty of fixing that
too. See the deleted assert_invalid in test/labels.wast and the added
$br_if1 testcase. `br_if` either sends its value to the destination label
or to its own result.