Batch of additional tests (#287)
For starters, more tests for:
* block
* unreachable
* nop
* comments

Also introduced a bit of sugar for exports. You can now write an export name as part of the function def itself:
```
(func "func-name" $func-name ...)
```
or, if the symbolic name isn't needed, just
```
(func "func-name" ...)
```
avoiding the tedious triple repetition of the name like before.