Switch back to no-push-void validation.
Also, flesh out $arity handling. Also, make `br`, `br_table`, `return`, and
`unreachable` return void instead of any (effectively) for now; this is a
significant semantic difference from ml-proto, and will need further
discussion. On the upside, it eliminates the need for `any` in this spec,
and along with eliminating `void`, this significantly simplifies the
validation logic.