Implement tables & multiple memories (#316)
As of WebAssembly/design#682. Still misses import/export abilities.

This changes & extends the S-expression syntax for tables and memories in the following way:
```
<elem_type>: anyfunc
<table>:     (table <nat> <nat>? <elem_type>)
<memory>:    (memory <nat> <nat>?)
<elem>:      (elem <expr> <var>*)
<data>:      (data <expr> <string>*)
```
In particular, memory segments are no longer part of the `memory` definition, in anticipation of the ability to import memory. Same for tables. This also mirrors the Wasm section structure more closely.

However, it is pretty tedious to count table elements in the common case, so the following shorthand is available:
```
(table <elem_type> (elem <var>*))    ;; = (table <size> <size> <elem_type>) (elem (i32.const 0) <var>*)
```
which pretty much behaves like the previous table syntax.
For symmetry, I introduced the analogous shorthand for memories, which turns out to be useful in tests as well:
```
(memory (data <string>*))   ;; = (memory <size> <size>) 
(data (i32.const 0) <string>*)
```
where `<size>` is the strings' total length in (rounded up) page units.

In the future, we can extend this syntax to name tables and memories, e.g.:
```
<table>:   (table <name>? <nat> <nat>? <elem_type>)
<memory>:  (memory <name>? <nat> <nat>?)
<elem>:    (elem <var>? <expr> <var>*)
<data>:    (data <var>? <expr> <string>*)
```