Type-check dead code and introduce assert_soft_invalid (#345)
We want to require valid code to be type-correct even in dead code, but allow engines to skip its validation. That is analogous to the liberty we give engines for validating individual functions lazily (which may include not at all if they are never called).

The motivation for this design choice is that typing rules should be simple but permissive enough and not require programmers or code generators to consider reachability. At the same time, it can be beneficial for implementations to be able to ignore dead code. The design represents a compromise: we only specify code to be valid if all instructions type-check, whether reachable or not. At the same time, we do not require (but still allow) engines to actually check the validity of code they can prove is never executed. The spec implements full checking, but the test suite distinguishes code that has to be identified as invalid by all engines (assert_invalid) and invalid code that they don't need to diagnose because it is unreachable (assert_soft_invalid).

This patch is going back to type-checking of dead code (with unconstrained initial stack), but respective tests use the new assert_soft_invalid assertion, which can be turned off for implementations that choose to skip dead code validation.