Generate tests that the Web Platform runner can pattern-match;
To determine that a file contains a testharness Web Platform test case, the
runner looks for an instance of a <script src=/resources/testharness.js>. So we
use this URI when generating the WPT test cases. In this mode, the testharness
files don't need to be generated, since the WPT server serves them.

The harness files aren't desired for the pure JS environment as well, since the
JS shell has to provide a shim mimicking the testharness.js APIs.