Give each test its own number, so that WPT expectation files can be automatically generated.
The WPT harness has a way to generate an expectation file specifying which
tests are known to be failing, and which ones are known to be passing. However,
to use this feature, it must have a unique identifier for each combination of
(file, test). To handle this, we generate a unique test number for each call to
test() in the wast harness.