Remove boolean types.
The WebAssembly community group was not able to reach a consensus on
whether boolean vector types should be included or not.

In order to move forward with prototyping of the proposal, remove the
boolean vector types along with supporting instructions. Change
comparison instructions to return a v128 mask vector and replace the
select instructions with a single v128.bitselect instruction.

These semantics are simply to implement on all current instruction set
architectures.