Fix invalid module names in test (#549)
The module names in the tests in elem.wast contained a "-", which is an invalid character in JavaScript variable names. Thereby the JavaScript test generated from elem.wast by the test script could not be executed anymore by JavaScript runtimes. This CL removes the "-" from the module names and thereby fixes the problem.