Editorial: Use USVString rather than DOMString to explain encoding
WebAssembly names are in the AST as strings of Unicode code points,
rather than UTF-8-encoded byte sequences. This patch uses WebIDL's
USVString type to encode this--the type is sequences of code points
in IDL-land, which is then converted to an ECMAScript value by
encoding in UTF-16. The place where this is inapplicable is custom
sections, which don't have a decoded name, but whose name is still
validated as UTF-8.

Thanks to @lukewagner for pointing this out in
https://github.com/WebAssembly/spec/pull/591#discussion_r151427259