Editorial: Clean up caching JS objects
Previously, the instantiate function had some kind of complicated logic
to make it so just one JS object per memory, table or exported function
would be created per address in the store without having a global mapping.
This patch switches to a global mapping to make the specification
easier to read and modify over time.

The change here was discussed in this comment thread:
https://github.com/WebAssembly/spec/pull/591#discussion_r151598115