Normative: Re-exported host functions are numbered
This patch makes the names of exported functions which come from
imported host functions be the index of the import. Although it
was a bit implicit in the original JS.md specification, this are
probably the originally indended behavior.

These semantics seem to be implemented by V8 and SpiderMonkey, but
JSC supports different semantics for exported function names.