[spec] Normative: Prohibit IndexedDB serialization of Modules
Based on discussions in the WebAssembly CG, it sounds like IndexedDB
serialization of WebAssembly Modules has the following problems:
- On some engines, it's impractical to serialize more than just the
  binary module data
- On some engines, upgrading from one version to another would cause
  the serialized module to be invalidated, with complicated implementation
  logic and a loss of reliability
- Some engines plan to just return an error for IndexedDB module
  serialization as a design decision

The purpose of IndexedDB Module serialization is for reliable performance,
and these items provide risks to that. This patch proposes that, initially,
IndexedDB serialization would not be provided for WebAssembly modules.