Potential copy&paste error in loop's note
This appears to me like a copy&paste error (coming from `block`'s note). You don't actually use this notation here. You use a simpler version with the empty label type. While this is not wrong as written (for `t^?` is allowed to be empty) it is confusing. But feel free to close if I am mistaken.