Rewrite the embedding limits test.
It did not work as written, and review comments were not addressed before
landing. A fair amount of code explicitly marked as "DO NOT COMMIT" was also
committed.

This commit makes the test use the `.any.js` feature that allows integration
with wpt and automated test suited. It also removes the unnecessary
abstraction of the run_* functions, ensures that the promise_test() tests
actually work, removes unactionable TODO comments, removes calls to print()
that break the test in a browser context, and uses the testharness.js library
much more idiomatically. The test now also uses `const` instead of `let` where
appropriate.