Stop trying to build js-api tests.
The build tool can't handle the current js-api tests, and would take a
significant rewrite to do so. In particular, it would need to stop handling the
helper .js files as tests, and ensure they are included when running the tests
that need them.

However, all of this is already achieved without a build step in the wpt context
as well as in the integration into SpiderMonkey's jstests harness. This
suggests that trying to handle these tests in the build script is not
necessary.