Fix bikeshed build
A recent change caused the generated html to have a sequence that looked
like `{{abkasdfadsf}}`, which bikeshed misinterprets as a WebIDL markup
shorthand. Since we never use that shorthand in the core spec, we can
disable it.

Since bikeshed thought this was a IDL reference, it inserted an
additional `<code>` tag, so the `mathjax2katex.py` script was not able
to extract the math blocks properly. The `FindMatching` function was
raising an `IndexError`, but it was never caught and instead terminated
the worker thread. The main thread would then wait forever for a worker
thread that will never finish. The fix here is to catch all exceptions
and print an error, instead of just catching `KeyboardInterrupt` and
`AssertionError`.