Update 64x2 instructions
A poll was held at CG meeting today, the consensus was for Option 3 of
https://github.com/WebAssembly/simd/issues/101.

This means we have all the f64x2 instructions, and some of the more
common i64x2 that have been benchmarked. This change reflects that
option in the proposal text, binary text, and implementation status.

Fixes #101.