Change some malformed shuffles to text parser errors
The tests here seem to be for the text format either specifying too few
or too few lanes, so this moves them to `assert_malformed` instead of
inline as an expected-valid module. This way parsers don't need to
handle a variable number of lanes but can always expect 16 lanes to be
specified.