[test] Get the value of a WebAssembly.Global in JS test_harness
The following test was failing in SpiderMonkey:

(module $Global
  (export "e" (global $g))
  (global $g i32 (i32.const 42))
)
(assert_return (get "e") (i32.const 42))

The issue was a generated call to index.js `get` which would return the
exported WebAssembly.Global object and compare it to a literal 42.

The harness emitted by js.ml seems to include this check. Not sure why
that's a separate harness, but for now it seems good to have the check
in all harnesses.