Update `ref.null` encoding after subtyping change (#149)
Issue https://github.com/WebAssembly/reference-types/issues/69 requires
that `ref.null` instructions include a reference type immediate. This
concept isn't present in the bulk-memory proposal, but the encoding is
(in element segment expressions).

This change updates the binary and text format, but not the syntax. This
is OK for now, since the only reference type allowed here is `funcref`.