Add .ocamlformat file to help format code
This file could be useful for autoformatting code. I would like to start
using it in some of the SIMD work, and it might be useful to have in in
the spec repo.

Note that this is not going to cause any automatic formatting, you would
have to run ocamlformat yourself. I also don't intend to reformat all
the code, since that will cause troublesome merge conflicts. I will
however like to format files that only exist in the SIMD proposal, which
will not cause a problem when we eventually merge it in.