[spec] Inconsistency of implementation-defined limits (#1195)
This PR resolves an inconsistency of implementation-defined limits in the js-api specification, and execution limits in the core specification. The core specification requires in https://webassembly.github.io/spec/core/appendix/implementation.html#execution that implementations throw a runtime error if tables or memories grow beyond an implementation-defined limit. The js-api specification, however, required a compile error if the module defines limits beyond the implementation-defined limits.

This PR resolves this inconsistency by adjusting the js-api specification. For the maximum memory size and maximum table size, a RuntimeError is now required instead of a CompileError.