Properly handle i32.trunc_f64_s of -2147483648.x (#1225)
Fixes #1224

Add testing for other values that seem too big to be
converted, but are ok after rounding. This is the only
one that was a spec interpreter bug. I wrote those
tests in a simple way with .9 values, replacing
the .9999999 from earlier - as @binji said, we should
either use hex-float notation, or avoid testing the parser
at the same time. As the issue here is rounding, I think
it's clearest to add .9 which suggests "this should be
rounded away".