Add SIMD type and value (#269)
Packed data is not defined anywhere. We can specify SIMD as a i128
(128-bit integer). Later when specifying the SIMD instructions, we can
use some "expansions" to expand i128 into the required lane shapes.

Also, use v128 instead of s128, v128 is already commonly used in many
places in this proposal, in the BinarySIMD.md, SIMD.md, and reference
interpreter.