Update label to syntax conventions
When you click on one of these links, you jump to the "Numeric
Instructions" section. But the definition of unop, binop, testop, relop,
and cvtop are in the "Conventions" section.

Move the label for these grouping non-terminals approriately, so
clicking on one of those definitions will go to where there are
defined.