Upgrade Sphinx to 3.5.2
The html_add_permalinks configuration was changed in 3.5
(https://www.sphinx-doc.org/en/3.x/usage/configuration.html#confval-html_add_permalinks),
we don't want any permalink for the heading (that's the existing
behavior).

Fix up bikeshed_fixup.py, we were not correctly replacing the spurious
"navigation section" and also the Unicode reference.

Most of the changes in the output is:

- adding viewport meta tag (in Sphinx#7695, since 3.1.0)
- sorting order of css files in html
- upgrade mathjax to 2.7.7