Fix quadratic behavior in reference interpreter.
The reference interpreter defines a `step` function on the program
configuration. This configuration embeds the current Wasm callstack and
block (label) stack, representing the state of execution with a tree of
terms. At each `step`, execution advances by one transition, e.g. one
instruction or one control-flow transfer.

When a `step` occurs on a configuration that has nested callstack and
block frames, the interpreter recurses structurally into the
configuration, eventually stepping in the innermost scope, and then
returning and rebuilding the entire configuration to the root of the
value. Unfortunately, this means that for deeply nested execution, the
interpreter can take `O(n)` time (where `n` is the depth of the stack of
all call and block frames) to execute a single instruction.

In bytecodealliance/wasmtime#3186, we found (via fuzzing against the
reference interpreter) a simple program that causes quadratic behavior
in the interpreter, taking many seconds to run despite a relatively
modest number of steps (1000 recursive calls nested in 18 block
levels per call).

This change to the interpreter adds a `step_many` function that resolves
the quadratic behavior, executing the above test program nearly
instantly. The key idea is to step an inner configuration (e.g., within
a block scope or function scope) many times until it cannot be stepped
"locally" (without the outer context). This prevents the unnecessary
exit and reentry to a deep recursive level (and corresponding
reallocation of the entire configuration tree) on every instruction
step.