WIP try using sphinx to generate w3c document
This uses a multihtml Sphinx builder, together with custom HTML layout
(layout.html and page.html), to build a document that matches w3c
requirements.

Currently the toctree directive does not work right, so it is duplicated
in the side bar (https://github.com/sphinx-doc/sphinx/pull/9830). The
W3C styles also don't apply fully to the side bar (because the Sphinx
generated TOC doesn't have the right class name, but this might be fixed
by some JS).

layout.html is used for all pages, index + individual pages, so the
front matter is guarded by a check that the current page is index.
layout.html is rather large, most of it is the front matter copied from
https://github.com/WebAssembly/spec/issues/1386#issuecomment-957347678.
page.html is used by all the individual rst pages.

Right now all the configuration lives in a new folder, newhtml, but it
can presumably replace the singlehtml folder.