[interpreter] use a correct module_inst /w invoke
With an `empty_module_inst` the very first `Invoke` instruction would be
executed with an empty module instance before the handling of the
`Invoke` would grab the instance stored within the function.

Today this isn’t doesn’t produce a wrong result – execution of the
`invoke` does not depend on access to the “correct” module instance.
This is true by the virtue of `Invoke` instruction, and the `invoke`
function, referencing a `func_inst`. However, if they did reference a
`funcaddr` instead, as prescribed in the specification, having an access
to the correct module instance would be a hard requirment for the
correct operation.