[spec] Initial JS API formal spec for the GC proposal
This commit adds formal spec test for the JS API portion of the GC
proposal. The semantics is based on the "no-frills" JS API proposal in
the following document:

  https://docs.google.com/document/d/17hCQXOyeSgogpJ0I0wir4LRmdvu4l7Oca6e1NkbVN8M/

The text assumes that the Wasm GC formal spec will expose certain
operations to the embedder, such as `ref.cast`. Comments have been left
where references between specs will need to be fixed up.