[test] Add JS API tests for Exported GC Objects
Adds a test for the draft JS API spec from this PR:

  https://github.com/WebAssembly/gc/pull/352

This primarily tests the internal object methods described in the spec.
It doesn't yet test the ToWebAssemblyValue / ToJSValue conversions.

Also syncs a portion of the module builder code with V8 to allow
constructing a subset of GC-using modules.