[spec] Restrict to `funcref` in inline element segment abbreviation syntax
An element list like `func $f $g` is itself an abbreviation for `funcref
(ref.func $f) (ref.func $g)`, so when the element list begins with
`func`, the only valid reftype for the corresponding table is `funcref`.

We should therefore only apply the abbreviation for element lists
containing function indices if the table’s reftype is already `funcref`.