Extends Name Section w/ Type & Field (#415)
This PR updates the spec to add additional subsections for type and field within the custom name section. struct.get and struct.set are also updated to indicate that their field indices are of type fieldidx, instead of u32. As a result, the metavariable representation was updated from i (integers) to y (indices).

Closes #193