[build] Use os.sched_getaffinity only if available
This change un-breaks the build in macOS and Windows environments, by
only calling the Python os.sched_getaffinity() function where it’s
available (which seems to be only on Linux) — and where it’s not
available, using multiprocessing.cpu_count() instead.

Otherwise, the build fails in macOS and Windows environment, because the
document/core/util/mathjax2katex.py program fails — and fails the build —
with a “module 'os' has no attribute 'sched_getaffinity'” message
(because os.sched_getaffinity(0) is hardcoded into the program).