[spec] Clarify binary element bitfield meaning.
In the binary format, when treating the number distinguishing different
kinds of element sections as a bitfield, the first bit distinguishes
active segements from passive/declarative segments.

The previous phrasing used "indicate" to mean that a 1 bit signifies
"passive or declarative", however it is easy to misread this as passive
and declarative referring to the two possible values of the bit instead.

This change modifies the description of Bit 0 to make it clearer that it
is differentiating active from passive or declarative, not passive from
declarative, by using the "distinguishes ... from ..." language used in
describing bit 1.