[js-api] Add tests for WebAssembly.JSTag (#319)
This adds tests for `WebAssembly.JSTag`. The tests are adapted from
https://github.com/v8/v8/blob/main/test/mjsunit/wasm/exnref-api.js.

Confirmed that they run successfully with web test infrastructure in
Chrome with `--js-flags=--experimental-wasm-exnref` command line
argument. The only thing failing there was the shadowrealm test, which I
think we should add a separate expectation files like the existing
`***.tentative.any.shadowrealm-expected.txt` in
https://github.com/chromium/chromium/tree/main/third_party/blink/web_tests/external/wpt/wasm/jsapi/exception.
But given that we don't host these files in this EH repo, I'll just
upload the js files.