Update Makefiles for easier generation of different document status
This is derived from #1764 but does not actually change the publication
status by default locally or on CI (ED is the default, and the w3c echidna
publication action overrides it to WD).
It also includes a fix to the echidna Makefile rules to exit with an error
in case an underlying step returns an error.