Add input for workflow dispatch for W3C publish workflow (#1791)
When manually dispatching workflows, inputs can be collected from the
dispatching user and passed to the workflow. Use this feature to
allow manually setting the W3C_STATE of the document. This will
let us publish CR drafts on every push (once we update the default
state from WD) and CR snapshots on demand.


Co-authored-by: Ms2ger <Ms2ger@igalia.com>