Update embedding and execution specs for memory64 (#74)
* Convert table and memory instance storage to sequences

Vectors in the spec are syntactically limited to a maximum size of 2^32.
This is incompatible with memory64, so the storage has been changed to
use sequences instead of vectors. The execution semantics are
unaffected.

* Update memory/table validation/execution/embedding to u64

Memory and table operations in the embedding section have been updated
to use u64 instead of u32, reflecting the changes to `limits`. Various
operations in the execution section were also updated to reflect
`idxtype`'s inclusion in `memtype` and `tabletype`, as well as the
supporting validation rules.