Expand tests for `array.new_data`
Noticed that the existing tests didn't exercise out-of-bounds indexing of the
data segment, so I added a test for that. At the same time, since I was already
creating a dedicated `.wast` file for testing `array.new_data`, I also added a
few general tests to round the file out. Some of these might technically
duplicate things that already happen to be tested in `array.wast`, but I think
that's probably fine, and that it is good to have dedicated tests in this new
file. I can remove them if we'd rather not have these potential-duplicate tests.