Use WPT version of test harness for HTML core test conversion
When converting tests to HTML for use in WPT, use the path to
the test harness in WPT's resources/ directory. This is also
important because the WPT test runner users this path inclusion
to discover the tests.

There are several more things I want to simplify here but I'm
starting with just this because it matches the sed transform
that @past originally wrote in
https://github.com/web-platform-tests/wpt/pull/49277