Separate tests for memory-related proposals into directories/files (#1964)
Based on the discussion in #1950,
splits tests for bulk memory, multi-memory, and memory64 (including test
generators) into separate directories and files. bulk memory is considered
an "artificial" dependency of the other 2 proposals, so for example tests which
cover the intersection between bulk memory and memory64 go into the
memory64 directory.